refactor: 파일 읽어들이는 로직 패키지 이동
